python/requirements.txt
openstacksdk
python-openstackclient
python-ironicclient
python-dracclient>=3.1.0
sushy>=2.0.0
redfishtool
jinja2
